Added a simple implementation of mail_queue.
The mail queue is a list at this moment. Might get extended in the future.

Also added clearing of the mail_queue to test setup and an error if the mail_queue is not empty
to test tear down. All tests are expected to check for any emails they send and clear them. Any
uncleared emails are probably there by mistake.

While doing this, I also noticed and fixed a small bug where send_email method was checking the wrong
config parameter.