Added page create, view and edit actions.
The controller is not finished yet. I am hoping it will be never used in its base form
but rather extended by classes such as GroupPageController. For that I created several hook methods
to facilitate easy processing of the objects.

The controller has no functional tests - not sure if it will need
those, but did not have enough time to write them anyway.

Feel free to rewrite it or send me comments if you see something very wrong with my chosen method.