Made subjects identifiable by location tag + id.
Refactored the way subjects are accessed.

Added routes that allow up to 5 levels of nesting for location tags in
subject url.

I had to disable file and page imports for subjects, because the
exports did not have university and faculty ids in them.

Notes:

We have subjects with identical ids in different faculties (at least
one named "anatomija").

Now there is an example of a list processing for formencode.

Also - apparently if you are using formencode, form_result should be
used, not request.POST or request.GET.