Added tags widget.
Based on the autobox jquery plugin. The plugin was modified to:
a) correctly fetch the value of suggestion returned via json
b) store tag names in the original input box too (to allow formencode to repopulate it if submission fails)

The theming of the widget was not modified yet. Waiting for final design.