Fixed some tests that were "sometimes" failing.
Ok, so apparently session variables are "session" variables which
means that after commit you might get another session from the
database.

The active user id is a session variable search configuration is a
session variable too, so it must be set after every commit to be safe.

We'll have to think of some convenient system of separating sections
in our tests, in functional tests controller does everything for you,
not so in unit tests, so we need to divide unit tests into some kind
of separate "actions".