Fixed file delete button.
Changed the moderator crowd admin crowd check - if its context is a file, it
checks if the user is an admin of the file's parent.

Also added a property to the file class, called file_parent. This property links files to
their subjects or groups. Should be done by SA in the future, using the content_item.parent.

Changed the cursor of the delete button to be a pointer.