Updated search - dealing with mixed tags now.
Mixed tags are simply tag names that map to both location tags and simple tags. That is, if
there are object's, whose location is 'some university' and objects simply tagged with 'some university',
a query with 'some university' in the tags section should return all those objects.

To achieve this, mixed tags are treated as a separate case.