Fixed gadugadu recipient bug.
The query to find all gadugadu recipients watching a subject had a glitch that meant it was ignoring whether the
user is GaduGadu-reachable. Fixed that.

Also added a trigger to maintain the sanity of gadugadu data: if the user has no uin, his other gadugadu fields are set to their
defaults.