Fixed issues with rendering the urls of deleted files.
Since deleting a file sets its parent to None, generating the url is a bit tricky.
Not to mention the fact that we do not want to show urls to deleted files.

At the moment links to deleted files are rendered as simple text. Later moderators should
be able to see these links.